refactor(sound): split SoundGenerator into category files
Mirrors the existing systems/sprites + SpriteGenerator pattern: each
SFX generator now lives in src/systems/sounds/ (footsteps, ui, combat,
quiz, movement) with shared WAV utilities in sounds/wav.ts.
SoundGenerator.ts remains the composition root and keeps its public API
(generateSounds, loadWav, encodeWAV) unchanged.

Addresses Tier C #5 from docs/architecture-review.md.
